Fix `Skills` icons
