+++ package.json
@@ -51,7 +51,6 @@
     "react-dom": "^18.1.0",
     "react-github-btn": "^1.2.0",
     "react-helmet": "^6.1.0",
-    "react-reveal": "^1.2.2",
     "react-scroll": "^1.8.1",
     "react-tilt": "^0.1.4",
     "sass": "^1.55.0"
