+++ src/components/Contact/Contact.jsx
@@ -4,7 +4,7 @@ import { Container } from 'react-bootstrap';
 import { OutboundLink } from 'gatsby-plugin-gtag';
 // import { PopupWidget } from 'react-calendly';
 import PortfolioContext from '../../context/context';
-import Title from '../Title/Title';
+import Title from '../Title';
 
 const Contact = () => {
   const { contact } = useContext(PortfolioContext);
@@ -16,18 +16,12 @@ const Contact = () => {
         <Title title="Contact" />
         <Fade bottom duration={1000} delay={800} distance="30px">
           <div className="contact-wrapper">
-            <p className="contact-wrapper__text">
-              {cta || 'Would you like to work with me? Awesome!'}
-            </p>
+            <p className="contact-wrapper__text">{cta || 'Would you like to work with me? Awesome!'}</p>
             <OutboundLink
               target="_blank"
               rel="noopener noreferrer"
               className="cta-btn cta-btn--resume"
-              href={
-                email
-                  ? `mailto:${email}`
-                  : 'https://github.com/cobidev/react-simplefolio'
-              }
+              href={email ? `mailto:${email}` : 'https://github.com/cobidev/react-simplefolio'}
             >
               {btn || "Let's Talk"}
             </OutboundLink>
