+++ src/components/Image/ProjectImg.jsx
@@ -25,9 +25,7 @@ const ProjectImg = ({ filename, alt, imgClassname }) => {
         }
       `}
       render={(data) => {
-        const image = data.images.edges.find((n) =>
-          n.node.relativePath.includes(filename),
-        );
+        const image = data.images.edges.find((n) => n.node.relativePath.includes(filename));
 
         if (!image) return null;
 
@@ -46,7 +44,7 @@ const ProjectImg = ({ filename, alt, imgClassname }) => {
 ProjectImg.propTypes = {
   filename: PropTypes.string.isRequired,
   alt: PropTypes.string.isRequired,
-  imgClassname: PropTypes.string,
+  imgClassname: PropTypes.string | undefined,
 };
 
 ProjectImg.defaultProps = {
