+++ src/components/Title.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Title = ({ title }: { title: string }) => <h2 className="section-title">{title}</h2>;
+
+export default Title;
