+++ src/images/index.d.ts
@@ -1,3 +1,4 @@
+declare module '*.gif';
 declare module '*.jpg';
 declare module '*.png';
 declare module '*.svg';
